modules/etf_quant_machine/requirements.txt
pandas==2.3.3
numpy==2.3.3
pyyaml==6.0.3
scikit-learn==1.7.2
scipy==1.16.2
matplotlib==3.10.7
openpyxl==3.1.5
tabulate==0.9.0
